chore(react): refactor action components
