+++ packages/react/src/components/actions/SignUpButton/SignUpButton.tsx
@@ -17,7 +17,7 @@
  */
 
 import {FC, forwardRef, ForwardRefExoticComponent, MouseEvent, ReactElement, Ref, RefAttributes, useState} from 'react';
-import useAsgardeo from '../../hooks/useAsgardeo';
+import useAsgardeo from '../../../hooks/useAsgardeo';
 import BaseSignUpButton, {BaseSignUpButtonProps} from './BaseSignUpButton';
 
 /**
